feat: entry content layout for ai
