+++ apps/desktop/layer/renderer/src/modules/entry-column/Items/picture-item.tsx
@@ -12,6 +12,7 @@ import { memo, use, useEffect, useMemo, useState } from "react"
 import { useTranslation } from "react-i18next"
 
 import { SwipeMedia } from "~/components/ui/media/SwipeMedia"
+import { useFeature } from "~/hooks/biz/useFeature"
 import { useRouteParamsSelector } from "~/hooks/biz/useRouteParams"
 import { EntryContent } from "~/modules/entry-content/components/entry-content"
 import { useImageDimensions } from "~/store/image"
@@ -43,9 +44,7 @@ export function PictureItem({ entryId, entryPreview, translation }: UniversalIte
               isActive && "rounded-b-none",
             )}
             imgClassName="object-cover"
-            onPreview={(media, i) => {
-              previewMedia(media, i)
-            }}
+            onPreview={previewMedia}
           />
         ) : (
           <div className="center bg-material-medium text-text-secondary aspect-square w-full flex-col gap-1 rounded-md text-xs">
@@ -75,6 +74,7 @@ export const PictureWaterFallItem = memo(function PictureWaterFallItem({
     id: state.id,
   }))
 
+  const aiEnabled = useFeature("ai")
   const isActive = useRouteParamsSelector(({ entryId }) => entryId === entry?.id)
   const entryContent = useMemo(() => <EntryContent entryId={entryId} noMedia compact />, [entryId])
   const previewMedia = usePreviewMedia(entryContent)
@@ -118,7 +118,7 @@ export const PictureWaterFallItem = memo(function PictureWaterFallItem({
               )}
               proxySize={proxySize}
               imgClassName="object-cover"
-              onPreview={previewMedia}
+              onPreview={aiEnabled ? undefined : previewMedia}
             />
 
             <div className="z-[3] shrink-0 overflow-hidden rounded-b-md pb-1">
