+++ apps/desktop/layer/renderer/src/modules/entry-content/components/layouts/index.ts
@@ -1,0 +1,5 @@
+export { ArticleLayout } from "./ArticleLayout"
+export { getEntryContentLayout } from "./factory"
+export { PicturesLayout } from "./PicturesLayout"
+export { SocialMediaLayout } from "./SocialMediaLayout"
+export { VideosLayout } from "./VideosLayout"
