+++ apps/desktop/layer/renderer/src/modules/entry-content/components/layouts/shared/index.ts
@@ -1,0 +1,3 @@
+export { AuthorHeader } from "./AuthorHeader"
+export { ContentBody } from "./ContentBody"
+export { VideoPlayer } from "./VideoPlayer"
