+++ packages/internal/components/src/ui/scroll-area/ScrollArea.tsx
@@ -117,15 +117,21 @@ const Root = ({
   ref: forwardedRef,
   className,
   children,
+  flex,
   ...rest
 }: React.ComponentPropsWithoutRef<typeof ScrollAreaBase.Root> & {
   ref?: React.Ref<React.ElementRef<typeof ScrollAreaBase.Root> | null>
+  flex?: boolean
 }) => (
   <ScrollAreaBase.Root
     {...rest}
     scrollHideDelay={0}
     ref={forwardedRef}
-    className={cn("overflow-hidden", className)}
+    className={cn(
+      "overflow-hidden",
+      flex && "min-h-0", // Add explicit min-height for flex contexts
+      className,
+    )}
   >
     {children}
     <Corner />
@@ -168,11 +174,14 @@ export const ScrollArea = ({
   return (
     <ScrollElementContext value={viewportRef}>
       <ScrollElementEventsContext value={events}>
-        <Root className={rootClassName}>
+        <Root className={rootClassName} flex={flex}>
           <Viewport
             ref={setViewportRef}
             onWheel={stopWheelPropagation ? stopPropagation : undefined}
-            className={cn(flex ? "[&>div]:!flex [&>div]:!flex-col" : "", viewportClassName)}
+            className={cn(
+              flex && "[&>div]:!flex [&>div]:!min-h-0 [&>div]:!flex-col", // Add min-h-0 to flex children
+              viewportClassName,
+            )}
             mask={mask}
             asChild={asChild}
             onScroll={onScroll}
